pyproject.toml
[tool.poetry]
name = "carshare"
version = "0.1.0"
description = "A web based Car Share management platform."
authors = ["George Goldberg <george@appletreetechnology.com>"]
license = "AGPL Version 3"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0"
django-tailwind = "^3.0.1"
psycopg-binary = "^3.0.7"
black = "^22.10"
django-allauth = "^0.51.0"
django-crispy-forms = "^1.13.0"
Pillow = "^9.3.0"
requests = "^2.27.1"
psycopg2-binary = "^2.9.3"
stripe = "^5.0.0"
celery = "^5.2.7"
django-celery-results = "^2.3.1"
django-celery-beat = "^2.3.0"
django-polymorphic = "^3.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
